switch to static og-images
